+++ plugins/system/usb/src/usb_main.c
@@ -168,7 +168,6 @@ void usb_init(ora_lookup_fn_t ora_lookup_fn) {
     ora_setup_usb_fn_t setup_usb = ora_lookup_fn(ORA_ID_SETUP_USB);
     ora_enable_irq_fn_t enable_irq = ora_lookup_fn(ORA_ID_ENABLE_IRQ);
     ora_get_clkref_mhz_fn_t get_clkref_mhz = ora_lookup_fn(ORA_ID_GET_CLKREF_MHZ);
-    ora_get_runtime_info_fn_t get_runtime_info = ora_lookup_fn(ORA_ID_GET_RUNTIME_INFO);
     context.set_status_led = ora_lookup_fn(ORA_ID_SET_STATUS_LED);
     context.get_active_ram_slot = ora_lookup_fn(ORA_ID_GET_ACTIVE_RAM_SLOT);
     context.get_ram_slot_info = ora_lookup_fn(ORA_ID_GET_RAM_SLOT_INFO);
@@ -176,14 +175,6 @@ void usb_init(ora_lookup_fn_t ora_lookup_fn) {
     context.reprogram_ram_rom_slot = ora_lookup_fn(ORA_ID_REPROGRAM_RAM_ROM_SLOT);
     // Can't log until we have the log functions
     DEBUG("USB plugin started");
-
-    // Get firmware and runtime information
-    if (get_runtime_info != NULL) {
-        context.runtime = get_runtime_info();
-    } else {
-        ERR("Runtime info function not available");
-        context.runtime = NULL;
-    }
 
     // Set up USB.  tinyusb will register its own IRQ handler, using the API
     // functions we provide.